Initial refactoring commit
